legacy-imports/hedge-pilot/pyproject.toml
[project]
name = "redflag-hedge"
version = "1.0.0"
description = "레드플래그 헤징 봇 - 김치 프리미엄 차익거래"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "gate-api>=6.104.0",
    "requests>=2.31.0",
    "PyJWT>=2.8.0",
    "colorama>=0.4.6",
]

[project.scripts]
redflag-hedge = "main:main"

[tool.uv]
dev-dependencies = []
